LPS-65351 Move ServiceAccessQuota to exported package
