+++ modules/apps/foundation/portal-security/portal-security-service-access-quota-service/src/main/java/com/liferay/portal/security/service/access/quota/ServiceAccessQuota.java
@@ -12,7 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.security.service.access.quota.internal;
+package com.liferay.portal.security.service.access.quota;
+
+import com.liferay.portal.security.service.access.quota.internal.SAQMetricConfig;
 
 import java.util.List;
